+++ setup.php
@@ -11,7 +11,7 @@ use GlpiPlugin\Tasksmanager\Profile;
 use GlpiPlugin\Tasksmanager\TaskDashboard;
 use GlpiPlugin\Tasksmanager\Workflow;
 
-define('PLUGIN_TASKSMANAGER_VERSION', '1.5.10');
+define('PLUGIN_TASKSMANAGER_VERSION', '1.6.0');
 define('PLUGIN_TASKSMANAGER_MIN_GLPI_VERSION', '11.0.0');
 define('PLUGIN_TASKSMANAGER_MAX_GLPI_VERSION', '11.0.99');